feat: reorganize entire events
