+++ src/main/events/library/index.ts
@@ -1,0 +1,5 @@
+import { getDBLibraryItems } from "@main/model/library";
+
+export async function handleGetLibraryGames() {
+	return await getDBLibraryItems();
+}
