+++ src/main/events/queue/index.ts
@@ -1,0 +1,12 @@
+import { syncronizeQueue } from "@main/model/queue";
+import { BrowserWindow, IpcMainEvent, IpcMainInvokeEvent } from "electron";
+
+export async function handleUpdateQueueItems(queueItems: IpcMainEvent) {
+	for (const win of BrowserWindow.getAllWindows()) {
+		win.webContents.send("updateQueueItems", queueItems);
+	}
+}
+
+export async function handleGetCurrentQueueItems(_event: IpcMainInvokeEvent) {
+	return await syncronizeQueue();
+}
