+++ src/main/events/torrent/types.ts
@@ -1,0 +1,6 @@
+import type { JSONGame } from "@main/worker";
+
+export interface Source {
+	name: string;
+	downloads: JSONGame[];
+}
